+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/facade/SoraVideoChannel.kt
@@ -23,21 +23,23 @@ class SoraVideoChannel(
         private val signalingEndpoint: String,
         private val channelId:         String,
         private val signalingMetadata: String = "",
+        private val clientId:          String? = null,
         private val spotlight:         Int = 0,
         private var streamType:        SoraStreamType,
         private var videoEnabled:      Boolean = true,
-        var         videoWidth:        Int = SoraVideoOption.FrameSize.Portrait.VGA.x,
-        var         videoHeight:       Int = SoraVideoOption.FrameSize.Portrait.VGA.y,
-        var         videoFPS:          Int =  30,
-        var         videoCodec:        SoraVideoOption.Codec = SoraVideoOption.Codec.VP9,
-        var         audioCodec:        SoraAudioOption.Codec = SoraAudioOption.Codec.OPUS,
-        var         videoBitrate:      Int? = null,
-        private var needLocalRenderer: Boolean = true,
-        private var audioEnabled:      Boolean = true,
-        private var sdpSemantics:      PeerConnection.SdpSemantics =
+        private val videoWidth:        Int = SoraVideoOption.FrameSize.Portrait.VGA.x,
+        private val videoHeight:       Int = SoraVideoOption.FrameSize.Portrait.VGA.y,
+        private val videoFPS:          Int =  30,
+        private val fixedResolution:   Boolean = false,
+        private val videoCodec:        SoraVideoOption.Codec = SoraVideoOption.Codec.VP9,
+        private val audioCodec:        SoraAudioOption.Codec = SoraAudioOption.Codec.OPUS,
+        private val videoBitrate:      Int? = null,
+        private val needLocalRenderer: Boolean = true,
+        private val audioEnabled:      Boolean = true,
+        private val sdpSemantics:      PeerConnection.SdpSemantics =
                 PeerConnection.SdpSemantics.UNIFIED_PLAN,
-        private var capturerFactory:   CameraVideoCapturerFactory =
-                DefaultCameraVideoCapturerFactory(context),
+        private val capturerFactory:   CameraVideoCapturerFactory =
+                DefaultCameraVideoCapturerFactory(context, fixedResolution),
         private var listener:          Listener?
 ) {
 
@@ -211,7 +213,8 @@ class SoraVideoChannel(
                 channelId          = channelId,
                 signalingMetadata  = signalingMetadata,
                 mediaOption        = mediaOption,
-                listener           = channelListener
+                listener           = channelListener,
+                clientId           = clientId
         )
         mediaChannel!!.connect()
     }