+++ CHANGES.md
@@ -12,6 +12,10 @@
 
 ## develop
 
+### UPDATE
+
+- Video chat に解像度固定、client ID を指定する選択肢を追加した
+
 ## 1.8.0
 
 ### UPDATE